% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_deps_anova.R
\name{analysis_deps_anova}
\alias{analysis_deps_anova}
\title{Differential expression analysis using ANOVA}
\usage{
analysis_deps_anova(expr_data_frame, group, log2_label = FALSE,
  return_padjust = TRUE, adjust_method = "BH")
}
\arguments{
\item{expr_data_frame}{A data frame containing ID and quantification values.}

\item{group}{A factor representing experimental groups.}

\item{log2_label}{A boolean value for representing whether the value is logarithmic or not, the default is FALSE.}

\item{return_padjust}{A boolean value for representing whether or not the p value is adjusted, the default is TRUE.}

\item{adjust_method}{Method used to adjust the p-values for multiple testing. See p.adjust for the complete list of options, the default is "BH".}
}
\value{
A data frame containing ID, log2(FC) and p value.
}
\description{
Differential expression analysis using ANOVA
}
\examples{
\dontrun{
anova_result <- analysis_deps_anova(
  expr_data_frame, group, log2_label = FALSE,
  return_padjust = TRUE, adjust_method = 'BH'
)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
