% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_deps_limma.R
\name{analysis_deps_limma}
\alias{analysis_deps_limma}
\title{Differential expression analysis using limma.}
\usage{
analysis_deps_limma(expr_data_frame, group, comparison_factor,
  log2_label = FALSE, adjust_method = "BH")
}
\arguments{
\item{expr_data_frame}{A data frame containing ID and quantification value.}

\item{group}{A factor for representing groups.}

\item{comparison_factor}{A vector for comparison factor.}

\item{log2_label}{A boolean value for representing whether or not the value is logarithmic, the default is FALSE.}

\item{adjust_method}{method used to adjust the p-values for multiple testing. See p.adjust for the complete list of options, the default is "BH"}
}
\value{
A list containing results from limma analysis.
}
\description{
Differential expression analysis using limma.
}
\examples{
\dontrun{
limma_results_list = analysis_deps_limma(
  expr_data_frame, group, comparison_statement,
  log2_label = FALSE, adjust_method = 'BH'
)
}
}
\references{
Ritchie, M.E., Phipson, B., Wu, D., Hu, Y., Law, C.W., Shi, W., and Smyth, G.K. (2015). limma powers differential expression \
analyses for RNA-sequencing and microarray studies. Nucleic Acids Research 43(7), e47.
}
\author{
Dongdong Zhan and Mengsha Tong
}
