% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mea_input.R
\name{check_mea_input}
\alias{check_mea_input}
\title{Check input for motif enrichment analysis (mea)}
\usage{
check_mea_input(foreground, background, center)
}
\arguments{
\item{foreground}{A vector for AA sequences with fixed length as foreground input.}

\item{background}{A vector for AA sequences with fixed length as background input.}

\item{center}{A character for center of k-mer.}
}
\value{
A list passing check steps
}
\description{
Check input for motif enrichment analysis (mea)
}
\examples{
## The process needs to load data from PhosMap datasets stored into FTP server and perform large computation.
## It may take a few minutes.
if(FALSE){
ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/check_mea_input.RData"
load_data <- load_data_with_ftp(ftp_url, 'RData')
writeBin(load_data, "check_mea_input.RData")
load("check_mea_input.RData")

check_result_list <- check_mea_input(
  foreground[1:100],
  background[1:100],
  center
)
head(check_result_list)
}

}
\author{
Dongdong Zhan and Mengsha Tong
}
