% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_kses.R
\name{compute_kses}
\alias{compute_kses}
\title{computing kinase-substrate enrichment score}
\usage{
compute_kses(
  substate_vector,
  regulons_of_kinase,
  substrates_of_kinase_in_exp_count
)
}
\arguments{
\item{substate_vector}{a vector for substrates with values indentified in current experiments.}

\item{regulons_of_kinase}{a vector for substrates of a specific kinase, which with substrates identified in current experiments.}

\item{substrates_of_kinase_in_exp_count}{a numeric for numbers in regulons_of_kinase vector.}
}
\value{
A numeric or NA for enrichment_score.
}
\description{
computing kinase-substrate enrichment score
}
\examples{
## The process needs to load data from PhosMap datasets stored into FTP server and perform large computation.
## It may take a few minutes.
if(FALSE){
ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/compute_kses.RData"
load_data <- load_data_with_ftp(ftp_url, 'RData')
writeBin(load_data, "compute_kses.RData")
load("compute_kses.RData")

stochastic_enrichment_score_i <- compute_kses(
  substate_vector,
  regulons_of_kinase_i,
  substrates_of_kinase_in_exp_count
)
head(stochastic_enrichment_score_i)
}
}
\references{
Hernandez-Armenta C et al. Benchmarking substrate-based kinase activity inference using phosphoproteomic data[J]. Bioinformatics, 2017, 33(12): 1845-1851.
}
\author{
Dongdong Zhan and Mengsha Tong
}
