% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fore_seq_to_motif.R
\name{fore_seq_to_motif}
\alias{fore_seq_to_motif}
\title{Convert the list that consists of motifs and the corresponding sequences to data frame.}
\usage{
fore_seq_to_motif(foreground_sequences_mapped_to_motifs)
}
\arguments{
\item{foreground_sequences_mapped_to_motifs}{A list that consists of motifs and the corresponding sequences.}
}
\value{
A data frame that consist of aligned sequences and the corresponding motifs.
}
\description{
Convert the list that consists of motifs and the corresponding sequences to data frame.
}
\examples{
ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/foreground_sequences_mapped_to_motifs.RData"
load_data <- load_data_with_ftp(ftp_url, 'RData')
writeBin(load_data, "foreground_sequences_mapped_to_motifs.RData")
load("foreground_sequences_mapped_to_motifs.RData")

df <- fore_seq_to_motif(
  foreground_sequences_mapped_to_motifs
)
head(df)

}
\author{
Dongdong Zhan and Mengsha Tong
}
