% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreground_sequences_mapped_to_motifs_list2df.R
\name{foreground_sequences_mapped_to_motifs_list2df}
\alias{foreground_sequences_mapped_to_motifs_list2df}
\title{Convert the list that consists of motifs and the corresponding sequences to data frame.}
\usage{

  foreground_sequences_mapped_to_motifs_list2df(foreground_sequences_mapped_to_motifs)
}
\arguments{
\item{foreground_sequences_mapped_to_motifs}{A list that consists of motifs and the corresponding sequences.}
}
\value{
A data frame that consist of aligned sequences and the corresponding motifs.
}
\description{
Convert the list that consists of motifs and the corresponding sequences to data frame.
}
\examples{
\dontrun{
df <- foreground_sequences_mapped_to_motifs_list2df(
  foreground_sequences_mapped_to_motifs
)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
