% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatted_output_mef_results.R
\name{formatted_output_mef_results}
\alias{formatted_output_mef_results}
\title{Output formatted sequences in foreground that are mapped to specific motifs.}
\usage{
formatted_output_mef_results(foreground_sequences_mapped_to_motifs)
}
\arguments{
\item{foreground_sequences_mapped_to_motifs}{A list that consists of motifs and their corresponding aligned sequences from foreground.}
}
\value{
A data frame that motifs and their corresponding aligned sequences from foreground.
}
\description{
Output formatted sequences in foreground that are mapped to specific motifs.
}
\examples{
\dontrun{
ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/formatted_output_mef_results.RData"
load_data <- load_data_with_ftp(ftp_url, 'RData')
writeBin(load_data, "formatted_output_mef_results.RData")
load("formatted_output_mef_results.RData")

formatted_output_df <- formatted_output_mef_results(
  foreground_sequences_mapped_to_motifs
)
head(formatted_output_df)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
