% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aligned_seq_for_mea.R
\name{get_aligned_seq_for_mea}
\alias{get_aligned_seq_for_mea}
\title{Taking S/T/Y as the center, align sequence to fasta library by specific length.}
\usage{
get_aligned_seq_for_mea(ID, Sequence, AA_in_protein, fixed_length,
  species = "human", fasta_type = "refseq")
}
\arguments{
\item{ID}{A vector for protein gi.}

\item{Sequence}{A vector for sequence of peptide.}

\item{AA_in_protein}{A vector for the location of S/T/Y in sequence of protein.}

\item{fixed_length}{A numeric value for aligned sequence,the default is 15.}

\item{species}{A string for that the alignment is based on which species, the options are human, mouse and rattus, the default is human.}

\item{fasta_type, }{A string for fasta source, the options are refseq and uniprot, the default is refseq}
}
\value{
A data frame containing ID, Sequence, AA_in_protein, aligned_seq.
}
\description{
Taking S/T/Y as the center, align sequence to fasta library by specific length.
}
\examples{
\dontrun{
aligned_sequence_df_based_on_fasta_library =
 get_aligned_seq_for_mea(
   ID,
   Sequence,
   AA_in_protein,
   fixed_length,
   species,
   fasta_type
 )
}
}
\references{
Hadley Wickham (2018). stringr: Simple, Consistent Wrappers for Common String Operations. R package version 1.3.0.\
https://CRAN.R-project.org/package=stringr.
}
\author{
Dongdong Zhan and Mengsha Tong
}
