% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colors_for_discrete_value.R
\name{get_colors_for_discrete_value}
\alias{get_colors_for_discrete_value}
\title{Generate custom colors from discrete values for heatmap.}
\usage{
get_colors_for_discrete_value(color_intervals_list, value_intervals_list)
}
\arguments{
\item{color_intervals_list}{a list for building color intervals.}

\item{value_intervals_list}{a list for building value intervals.}
}
\value{
A vectors containing colors distribution.
}
\description{
Generate custom colors from discrete values for heatmap.
}
\examples{
\dontrun{
value_intervals_list = list(
seq(-4, -2, 0.2),
seq(-2, -1, 0.2),
seq(-1, 1, 0.2),
seq(1, 2, 0.2),
seq(2, 4, 0.2)
)
color_intervals_list = list(
  c('blue', '#33CCFF'),
  c('#33CCFF', 'green'),
  c('green', 'white', '#FF6600'),
  c('#FF6600', 'red'),
  c('red', 'firebrick')
)
colors = get_colors_for_discrete_value(
  color_intervals_list,
  value_intervals_list
)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
