% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_combined_data_frame_with_mapped_gene_symbol.R
\name{get_combined_data_frame_with_mapped_gene_symbol}
\alias{get_combined_data_frame_with_mapped_gene_symbol}
\title{Get a data frame mapped GI number to Gene Symbol.}
\usage{
get_combined_data_frame_with_mapped_gene_symbol(merge_df_with_phospho_peptides,
  species = "human")
}
\arguments{
\item{merge_df_with_phospho_peptides}{A dataframe consisting of IDs (Sequence_GI_Psite) and Area values.}

\item{species}{A string, the options is human and mouse, the default is human.}
}
\value{
A dataframe with Sequence, GI, Modification, Gene Symbol, Area values and PSMs
}
\description{
This is an intermediate file and a dataframe with Gene Symbol is exported.
Based on a library file consisting of mapping relationships about Gene Symbol, GeneID and GI,
a new dataframe with Sequence, GI, Modification, Gene Symbol, Area and PSMs,is contructed.
}
\examples{
\dontrun{
combinated_df_with_mapped_gene_symbol = get_combined_data_frame_with_mapped_gene_symbol(
  merge_df_with_phospho_peptides
)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
