% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_with_AAs_i_of_reducing_redundancy.R
\name{get_df_with_AAs_i_of_reducing_redundancy}
\alias{get_df_with_AAs_i_of_reducing_redundancy}
\title{Get data frame without redundancy.}
\usage{
get_df_with_AAs_i_of_reducing_redundancy(df_with_AAs_i)
}
\arguments{
\item{df_with_AAs_i}{a data frame for peptides of the ith protein.}
}
\value{
A data frame with sites in unique protein.
}
\description{
Get data frame without redundancy.
}
\author{
Dongdong Zhan and Mengsha Tong
}
