% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_info_from_dir.R
\name{get_file_info_from_dir}
\alias{get_file_info_from_dir}
\title{Get data lists from and corresponding file ids.}
\usage{
get_file_info_from_dir(specific_dir, experiment_ID)
}
\arguments{
\item{specific_dir}{A folder containing files as input.}

\item{experiment_ID}{A vector containing experiment codes as input}
}
\value{
A list containing data from files and corresponding file ids
}
\description{
Read batch files (.txt or .csv) from specific directory.
}
\examples{
\dontrun{
result_list = get_file_info_from_dir(
  specific_dir,
  experiment_ID
)
}
}
