% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filtered_df.R
\name{get_filtered_df}
\alias{get_filtered_df}
\title{Get data frame filtered based on the Mascot and reference files.}
\usage{
get_filtered_df(mascotfileName, refFileName)
}
\arguments{
\item{mascotfileName}{a string for mascot names as input.}

\item{refFileName}{a string for reference file names.}
}
\value{
A filtered data frame
}
\description{
Get data frame filtered based on the Mascot and reference files.
}
\examples{
\dontrun{
df <- get_filtered_df(mascotfileName, refFileName)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
