% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_foreground_df_to_motifs.R
\name{get_foreground_df_to_motifs}
\alias{get_foreground_df_to_motifs}
\title{Get filtered foreground data frame that its aligned sequences with specific motif.}
\usage{
get_foreground_df_to_motifs(
  foreground_sequences_mapped_to_motifs,
  foreground,
  foreground_df
)
}
\arguments{
\item{foreground_sequences_mapped_to_motifs}{A list that consists of motifs and its corresponding aligned sequences.}

\item{foreground}{A vector for aligned sequences.}

\item{foreground_df}{A data frame from the initial foreground data frame.}
}
\value{
A data frame that its aligned sequences with specific motif.
}
\description{
Get filtered foreground data frame that its aligned sequences with specific motif.
}
\examples{
## The process needs to load data from PhosMap datasets stored into FTP server and perform large computation.
## It may take a few minutes.
if(FALSE){
    ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/get_foreground_df_to_motifs.RData"
    load_data <- load_data_with_ftp(ftp_url, 'RData')
    writeBin(load_data, "get_foreground_df_to_motifs.RData")
    load("get_foreground_df_to_motifs.RData")

    foreground_df_mapped_to_motifs <- get_foreground_df_to_motifs(
      foreground_sequences_mapped_to_motifs,
      foreground, foreground_df
    )
   head(foreground_df_mapped_to_motifs)
}


}
\author{
Dongdong Zhan and Mengsha Tong
}
