% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_background_df.R
\name{get_global_background_df}
\alias{get_global_background_df}
\title{Get background data frame (fasta library from Refseq).}
\usage{
get_global_background_df(species = "human", fasta_type = "refseq")
}
\arguments{
\item{species}{A string for that the alignment is based on which kinds of species, the options are human, mouse and rattus.}

\item{fasta_type, }{A string for fasta source, the options are refseq and uniprot, the default is refseq}
}
\value{
A data frame of background
}
\description{
Get background data frame (fasta library from Refseq).
}
\examples{
\dontrun{
background_df <- get_global_background_df(species)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
