% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ksea_result_list.R
\name{get_ksea_result_list}
\alias{get_ksea_result_list}
\title{Kinase activity analysis based on known and predicted kinase-substrate relationships}
\usage{
get_ksea_result_list(ptypes_data_ratio_in_single_exp, ID,
  kinase_substrate_regulation_relationship, ksea_activity_i_pvalue = 0.05)
}
\arguments{
\item{ptypes_data_ratio_in_single_exp}{A quantification vector from a single experiment.}

\item{ID}{A phosporylation ID vector like VIM_S56 (GeneSymbol_psite).}

\item{kinase_substrate_regulation_relationship}{A data frame contanning kinase-substrate relationships that consists of "kinase", "substrate", "site", "sequence" and "predicted" columns.}

\item{ksea_activity_i_pvalue}{A cutoff used for filtering significant activities computed from KSEA.}
}
\value{
A list containing results from ksea.
}
\description{
Kinase activity analysis based on known and predicted kinase-substrate relationships
}
\examples{
ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/get_ksea_result_list.RData"
load_data <- load_data_with_ftp(ftp_url, 'RData')
writeBin(load_data, "get_ksea_result_list.RData")
load("get_ksea_result_list.RData")

ksea_result_list_i <- get_ksea_result_list(
 ptypes_data_ratio_in_single_exp, ID,
 kinase_substrate_regulation_relationship,
 ksea_activity_i_pvalue = 0.05
)
head(ksea_result_list_i)
}
\author{
Dongdong Zhan and Mengsha Tong
}
