% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kses.R
\name{get_kses}
\alias{get_kses}
\title{computing kinase-substrate enrichment significance (pvalue)}
\usage{
get_kses(substate_vector, regulons_of_kinase, trial = 1000)
}
\arguments{
\item{substate_vector}{a vector for substrates with values identified in current experiments.}

\item{regulons_of_kinase}{a vector for substrates of a specific kinase, which identified in current experiments.}

\item{trial}{a numeric for the number of random samples, the default is 1000.}
}
\value{
A list for expected enrichment scores and its significance
}
\description{
computing kinase-substrate enrichment significance (pvalue)
}
\examples{
\dontrun{
pvalue <- get_kses(
  substate_vector,
  regulons_of_kinase,
  1000
)
}

}
\references{
Hernandez-Armenta C et al. Benchmarking substrate-based kinase activity inference using phosphoproteomic data[J]. Bioinformatics, 2017, 33(12): 1845-1851.
}
\author{
Dongdong Zhan and Mengsha Tong
}
