% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list_with_filted_sites.R
\name{get_list_with_filted_sites}
\alias{get_list_with_filted_sites}
\title{Filter phosphorylation sites.}
\usage{
get_list_with_filted_sites(peptide_id, files, files_site_score, qc, min_score,
  min_FDR)
}
\arguments{
\item{peptide_id}{A vector containing experiment ids as input.}

\item{files}{A data list containing peptides identificaton from Firmiana as input.}

\item{files_site_score}{A data list containing psites scores extracted from mascot xml. The default is NULL,
which represents no QC file.}

\item{qc}{A boolean value representing whether it has QC files. The default is True.}

\item{min_score}{A numeric for the minimum score of credible peptides, the default is 20 for Mascot ion score.}

\item{min_FDR}{A numeric for the minimum FDR of credible peptides, the default is 0.01.}
}
\value{
A list containing peptides dataframe with area values and psm,
IDs with mergered sequences, gi and sites, new peptides dataframe combined previous peptides dataframe and IDs.
}
\description{
Filter phosphorylation sites by extracting all peptides with ion score>=20 and FDR<0.01 from Firmiana and having psites scores.
}
\examples{
\dontrun{
result_list_with_filtered_sites = get_list_with_filted_sites(
  peptide.id,
  files,
  files_site_score
)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
