% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_merged_phospho_df.R
\name{get_merged_phospho_df}
\alias{get_merged_phospho_df}
\title{Get merged data frame with phospho-peptides.}
\usage{
get_merged_phospho_df(peptide_id, peptide_df_with_area_psm_list,
  ID_of_seq_gi_site_list, ID_DF_list)
}
\arguments{
\item{peptide_id}{a vector for peptide ID.}

\item{peptide_df_with_area_psm_list}{a list for peptides with areas and PSMs.}

\item{ID_of_seq_gi_site_list}{a list for peptides ID with sequence, gi and site.}

\item{ID_DF_list}{a list for ID and values.}
}
\value{
a merged data frame with phospho-peptides.
}
\description{
Get merged data frame with phospho-peptides.
}
