% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modification_index.R
\name{get_modification_index}
\alias{get_modification_index}
\title{Get index of modifications in protein sequence.}
\usage{
get_modification_index(id_data_only_peptide2gi, fasta_data)
}
\arguments{
\item{id_data_only_peptide2gi}{a data frame for peptides with protein gi.}

\item{fasta_data}{a fasta data for a specific species.}
}
\value{
A vector for index of modifications in protein sequence.
}
\description{
Get index of modifications in protein sequence.
}
\author{
Dongdong Zhan and Mengsha Tong
}
