% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modification_index_in_protein_seq_list.R
\name{get_modification_index_in_protein_seq_list}
\alias{get_modification_index_in_protein_seq_list}
\title{Get index of modifications in protein sequence.}
\usage{
get_modification_index_in_protein_seq_list(id_data_only_peptide2gi, gi_fasta)
}
\arguments{
\item{id_data_only_peptide2gi}{a data frame for peptides with protein gi.}

\item{gi_fasta}{a fasta data for a specific species.}
}
\value{
A vector for index of modifications in protein sequence.
}
\description{
Get index of modifications in protein sequence.
}
\author{
Dongdong Zhan and Mengsha Tong
}
