% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_motif_analysis_summary.R
\name{get_motif_analysis_summary}
\alias{get_motif_analysis_summary}
\title{Get summary results of motif analysis for specific input}
\usage{
get_motif_analysis_summary(
  foreground,
  background,
  center = "S",
  min_sequence_count = 1,
  min_pvalue = 0.01
)
}
\arguments{
\item{foreground}{A vector for AA sequences with fixed length as foreground input.}

\item{background}{A vector for AA sequences with fixed length as background input.}

\item{center}{A character for center of k-mer.}

\item{min_sequence_count}{A numeric for the minimum sequence number assigned to a motif.}

\item{min_pvalue}{A numeric for the minimum pvalue for found motif.}
}
\value{
A list for summary result of motif analysis
}
\description{
Get summary results of motif analysis for specific input
}
\examples{
## The process needs to load data from PhosMap datasets stored into FTP server and perform large computation.
## It may take a few minutes.
if(FALSE){
    ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/get_motif_analysis_summary.RData"
    load_data <- load_data_with_ftp(ftp_url, 'RData')
    writeBin(load_data, "get_motif_analysis_summary.RData")
    load("get_motif_analysis_summary.RData")

    motifs <- get_motif_analysis_summary(
      foreground[1:100], background[1:100],
      center = center,
      min_sequence_count = 1,
      min_pvalue = motifx_pvalue
    )
    head(motifs)
}

}
\references{
Omar Wagih (2014). rmotifx: An iterative statistical approach to the discovery of biological sequence motifs. R package version 1.0.
}
\author{
Dongdong Zhan and Mengsha Tong
}
