% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_motifs_list.R
\name{get_motifs_list}
\alias{get_motifs_list}
\title{Motif enrichment using rmotifx.}
\usage{
get_motifs_list(foreground, background, center_vector, motifx_pvalue)
}
\arguments{
\item{foreground}{A vector for aligned sequences as the foreground input.}

\item{background}{A vector for aligned sequences as the background input.}

\item{center_vector}{A vector for aligned centers.}

\item{motifx_pvalue}{A numeric value for selecting motifs that meets the minimum cutoff.}
}
\value{
A list for results of motif enrichment.
}
\description{
Motif enrichment using rmotifx.
}
\examples{
## The process needs to load data from PhosMap datasets stored into FTP server and perform large computation.
## It may take a few minutes.
if(FALSE){
    ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/get_motifs_list.RData"
    load_data <- load_data_with_ftp(ftp_url, 'RData')
    writeBin(load_data, "get_motifs_list.RData")
    load("get_motifs_list.RData")

    motifs_list <- get_motifs_list(foreground[1:100], background[1:100], center_vector, motifx_pvalue)
    head(motifs_list)
}
}
\references{
Omar Wagih (2014). rmotifx: An iterative statistical approach to the discovery of biological sequence motifs. R package version 1.0.
}
\author{
Dongdong Zhan and Mengsha Tong
}
