% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_normalized_data_FOT5.R
\name{get_normalized_data_FOT5}
\alias{get_normalized_data_FOT5}
\title{Normailization on basis of sum}
\usage{
get_normalized_data_FOT5(data_frame, experiment_code_file_path)
}
\arguments{
\item{data_frame}{A data frame containing IDs and values merged from multi-experiments as input.}

\item{experiment_code_file_path}{A file path of storing experiment codes as input. The experiment codes are required to keep pace with column names of Values.}
}
\value{
A data frame after normalization
}
\description{
Normailization on basis of sum
}
\examples{
demo_data_url1 <- url('https://raw.githubusercontent.com/ecnuzdd/PhosMap_datasets/master/function_demo_data/profiling_data.RData')
demo_data_url2 <- url('https://raw.githubusercontent.com/ecnuzdd/PhosMap_datasets/master/function_demo_data/profiling_exp_design_info.txt')
load(demo_data_url1)
profiling_exp_design_info_file_path <- demo_data_url2

profiling_data_normalized <- get_normalized_data_FOT5(profiling_data,
  profiling_exp_design_info_file_path)

head(profiling_data_normalized)

}
\author{
Dongdong Zhan and Mengsha Tong
}
