% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_df_of_unique_proteins_with_sites.R
\name{get_summary_df_of_unique_proteins_with_sites}
\alias{get_summary_df_of_unique_proteins_with_sites}
\title{Assign psites to protein sequence.}
\usage{

  get_summary_df_of_unique_proteins_with_sites(combinated_df_with_mapped_gene_symbol,
  species = "human")
}
\arguments{
\item{combinated_df_with_mapped_gene_symbol}{A dataframe with Sequence, GI, Modification, Gene Symbol, Area and PSMs as input.}

\item{species}{A string, the options is human and mouse, the default is human.}
}
\value{
A dataframe that all redundant psites are assigned to protein sequence.
}
\description{
Construct the data frame with unique phosphorylation site for each protein sequence and eliminate redundancy.
}
\examples{
\dontrun{
summary_df_of_unique_proteins_with_sites = get_summary_df_of_unique_proteins_with_sites(
  combinated_df_with_mapped_gene_symbol
)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
