% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_with_unique_sites.R
\name{get_summary_with_unique_sites}
\alias{get_summary_with_unique_sites}
\title{Assign psites to protein sequence.}
\usage{
get_summary_with_unique_sites(combined_df_with_mapped_gene_symbol,
  species = "human", fasta_type = "refseq")
}
\arguments{
\item{combined_df_with_mapped_gene_symbol}{A dataframe with Sequence, ID, Modification, Gene Symbol, Area and PSMs as input.}

\item{species}{A string, the options are human, mouse and rattus, the default is human.}

\item{fasta_type, }{A string for fasta source, the options are refseq and uniprot, the default is refseq}
}
\value{
A dataframe that all redundant psites are assigned to protein sequence.
}
\description{
Construct the data frame with unique phosphorylation site for each protein sequence and eliminate redundancy.
}
\examples{
demo_data_url <- url('https://raw.githubusercontent.com/ecnuzdd/PhosMap_datasets/master/function_demo_data/combined_df_with_mapped_gene_symbol.RData')
load(demo_data_url)

summary_df_of_unique_proteins_with_sites <- get_summary_with_unique_sites(
  combined_df_with_mapped_gene_symbol,
  species = 'human',
  fasta_type = 'refseq')

head(summary_df_of_unique_proteins_with_sites)

}
\author{
Dongdong Zhan and Mengsha Tong
}
