% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_AAs_i_df.R
\name{get_unique_AAs_i_df}
\alias{get_unique_AAs_i_df}
\title{Get data frame without redundancy.}
\usage{
get_unique_AAs_i_df(df_with_AAs_i)
}
\arguments{
\item{df_with_AAs_i}{a data frame for peptides of the ith protein.}
}
\value{
A data frame with sites in unique protein.
}
\description{
Get data frame without redundancy.
}
\examples{
\dontrun{
ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/get_unique_AAs_i_df.RData"
load_data <- load_data_with_ftp(ftp_url, 'RData')
writeBin(load_data, "get_unique_AAs_i_df.RData")
load("get_unique_AAs_i_df.RData")

summary_df_of_unique_protein_with_sites <- get_unique_AAs_i_df(df_with_AAs_i)
head(summary_df_of_unique_protein_with_sites)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
