% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello, world!}
\usage{
hello()
}
\description{
This is an example function named 'hello'
which prints 'Hello, world!' and 'Welcome to PhosMap!'.
}
\details{
Some useful keyboard shortcuts for package authoring:
Build and Reload Package:  'Ctrl + Shift + B'
Check Package:             'Ctrl + Shift + E'
Test Package:              'Ctrl + Shift + T'
}
\author{
Dongdong Zhan and Mengsha Tong
}
