% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_psites_with_max_in_topX.R
\name{keep_psites_with_max_in_topX}
\alias{keep_psites_with_max_in_topX}
\title{Keep psites whose row maximum is top N (percentage).}
\usage{
keep_psites_with_max_in_topX(phospho_data, percent_of_kept_sites = 3/4)
}
\arguments{
\item{phospho_data}{A data frame of phospho-data.}

\item{percent_of_kept_sites}{A numeric value representing a cutoff used for filter psites. The default is 3/4.}
}
\value{
A data frame meeting specific cutoff.
}
\description{
Compute row maximum each psites, sort row maximum in decreasing order and keep top N (percentage).
}
\examples{
demo_data_url <- url('https://raw.githubusercontent.com/ecnuzdd/PhosMap_datasets/master/function_demo_data/phospho_data.RData')
load(demo_data_url)

phospho_data_topX <- keep_psites_with_max_in_topX(phospho_data,
  percent_of_kept_sites = 0.9)

head(phospho_data_topX)

}
\author{
Dongdong Zhan and Mengsha Tong
}
