% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_with_ftp.R
\name{load_data_with_ftp}
\alias{load_data_with_ftp}
\title{Load datasets from URL (ftp://111.198.139.72:4000/pub/PhosMap_datasets)}
\usage{
load_data_with_ftp(ftp_link, data_type)
}
\arguments{
\item{ftp_link}{A string for URL of datasets.}

\item{data_type}{A string for type of datasets (txt, csv, RData).}
}
\value{
A dataframe
}
\description{
Some datasets with larger size need to be loaded for mapping ID or protein sequence when using PhosMap.
These datasets could be ragarded as library and uploaded to ftp://111.198.139.72:4000/pub/PhosMap_datasets in advance.
When first perfoming functions depending on these datasets, PhosMap will get them from specific URL and save them into local disk.
}
\examples{
ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/profiling_exp_design_info.txt"
load_data <- load_data_with_ftp(ftp_url, 'txt')
head(load_data)

}
\author{
Dongdong Zhan and Mengsha Tong
}
