% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_profiling_file_from_Firmiana.R
\name{merge_profiling_file_from_Firmiana}
\alias{merge_profiling_file_from_Firmiana}
\title{Merge profiling files downloaded from Firmiana.}
\usage{
merge_profiling_file_from_Firmiana(firmiana_gene_dir, US_cutoff = 1,
  experiment_gene_file_path)
}
\arguments{
\item{firmiana_gene_dir}{a folder containing gene identification results as input.}

\item{US_cutoff}{a numerical value as a cutoff to filter data, the default is 1.}

\item{experiment_gene_file_path}{a file path for storing experiemnt design of proteomics data.}
}
\value{
A merged data frame after filtering (US_cutoff) and replacing NAs to zeros.
}
\description{
Filter data based on US (Unique and Ionscore > 20) peptide counts then merge profiling files.
}
\examples{
\dontrun{
merged_df = merge_profiling_file_from_Firmiana(firmiana_gene_dir, US_cutoff = 1)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
