% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/motif_get_aligned_sequence_df_based_on_fasta_library.R
\name{motif_get_aligned_sequence_df_based_on_fasta_library}
\alias{motif_get_aligned_sequence_df_based_on_fasta_library}
\title{Taking S/T/Y as the center, align sequence to fasta library by specific length.}
\usage{
motif_get_aligned_sequence_df_based_on_fasta_library(GI, Sequence,
  AA_in_protein, fixed_length, species)
}
\arguments{
\item{GI}{A vector for protein gi.}

\item{Sequence}{A vector for sequence of peptide.}

\item{AA_in_protein}{A vector for the location of S/T/Y in sequence of protein.}

\item{fixed_length}{A numeric value for aligned sequence,the default is 15.}

\item{species}{A string for that the alignment is based on which species, the options are 'human' and 'mouse'}
}
\value{
A data frame containing GI, Sequence, AA_in_protein, aligned_seq.
}
\description{
Taking S/T/Y as the center, align sequence to fasta library by specific length.
}
\examples{
\dontrun{
aligned_sequence_df_based_on_fasta_library =
 motif_get_aligned_sequence_df_based_on_fasta_library(
   GI,
   Sequence,
   AA_in_protein,
   fixed_length,
   species
 )
}
}
\references{
Hadley Wickham (2018). stringr: Simple, Consistent Wrappers for Common String Operations. R package version 1.3.0.\
https://CRAN.R-project.org/package=stringr.
}
\author{
Dongdong Zhan and Mengsha Tong
}
