% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_based_on_proteomics_maxquant.R
\name{norm_based_on_proteomics_maxquant}
\alias{norm_based_on_proteomics_maxquant}
\title{Normalizaiton for phosphoproteomics data from MaxQuant based on proteomics data}
\usage{
norm_based_on_proteomics_maxquant(
  summary_phos_norm,
  proteomics_data,
  experiment_code_file_path,
  proteomics_experiment_file_path,
  intensity_type = "Intensity",
  min_unique_peptide = 1,
  max_na_num = 2,
  norm_method = "global",
  impute_method = "minimum/10"
)
}
\arguments{
\item{summary_phos_norm}{A data frame containing information required for all analysis.}

\item{proteomics_data}{A data frame of proteinGroups.txt.}

\item{experiment_code_file_path}{Experiment code file path for phosphoproteomics.}

\item{proteomics_experiment_file_path}{Experiment code file path for proteomics.}

\item{intensity_type}{Intensity type. The default is 'Intensity' and the options are 'iBAQ' and 'LFQ.intensity',depending on proteinGroups.txt.}

\item{min_unique_peptide}{Threshold for MaxQuant unique peptide[proteinGroups.txt]. The default is 1.}

\item{max_na_num}{Threshold for the number of missing values[proteinGroups.txt]. The default is 2.}

\item{norm_method}{Normalizaiton method[proteinGroups.txt]. The default is 'global' and the options are 'global' and 'median'.}

\item{impute_method}{Imputation method[proteinGroups.txt]. The default is 'minimum/10',the options are '0', 'minimum' and 'minimum/10'.}
}
\value{
A result list. Elements are a data frame containing information required for all analysis and a pre-processed proteomics data.
}
\description{
Normalizaiton for phosphoproteomics data from MaxQuant based on proteomics data
}
\examples{

## Read phosphoproteomics data
\dontrun{
rawdata <- read.csv("Phospho (STY)Sites.txt",header=T,sep='\t')

## Quality control for phosphoproteomics data 
qc_results <- qc_maxquant(rawdata, "./experiment_code_file.txt", min_score = 40, min_loc_prob = 0.75, max_na_num = 2)
qc_result <- qc_results[[1]]
qc_result_for_motifanalysis <- qc_results[[2]]

## Normalizaiton, imputation and filtering
summary_phos_norm <- norm_maxquant(qc_result, qc_result_for_motifanalysis, norm_method = "global", impute_method = "minimum/10", top = 0.9)

## Read proteomics data
proteomics_data <- read.csv("./proteinGroups.txt", sep = "\t")

results <- norm_based_on_proteomics(summary_phos_norm, proteomics_data, "./phosphorylation_exp_design_info.txt", "./phosphorylation_exp_design_info.txt")
summary_phos_norm_based_on_pro <- results[[1]]
pro_norm <- results[[2]]
}
}
