% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_maxquant.R
\name{norm_maxquant}
\alias{norm_maxquant}
\title{Normalizaiton, imputation and filtering for phosphoproteomics data from MaxQuant.}
\usage{
norm_maxquant(
  qc_result,
  qc_result_for_motifanalysis,
  norm_method = "global",
  impute_method = "minimum/10",
  percent_of_kept_sites = 3/4
)
}
\arguments{
\item{qc_result}{A data frame containing quality control result.}

\item{qc_result_for_motifanalysis}{A data frame containing information required for motif analysis.}

\item{norm_method}{Normalizaiton method. The default is 'global' and the options are 'global' and 'median'.}

\item{impute_method}{Imputation method. The default is 'minimum/10' and the options are '0', 'minimum' and 'minimum/10'.}

\item{percent_of_kept_sites}{A numeric value representing a cutoff used for filter psites. The default is 3/4.}
}
\value{
A data frame containing information required for all analysis.
}
\description{
Normalizaiton, imputation and filtering for phosphoproteomics data from MaxQuant.
}
\examples{
\dontrun{
rawdata <- read.csv("Phospho (STY)Sites.txt",header=T,sep='\t')

## Quality control for phosphoproteomics data 
qc_results <- qc_maxquant(rawdata, "./experiment_code_file.txt", min_score = 40, min_loc_prob = 0.75, max_na_num = 2)
qc_result <- qc_results[[1]]
qc_result_for_motifanalysis <- qc_results[[2]]

summary_phos_norm <- norm_maxquant(qc_result, qc_result_for_motifanalysis, norm_method = "global", impute_method = "minimum/10", top = 0.9)
}
}
