% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_nopair_ctrl_by_col.R
\name{normalize_nopair_ctrl_by_col}
\alias{normalize_nopair_ctrl_by_col}
\title{For data without pairs but with control, normalize them to the control.}
\usage{
normalize_nopair_ctrl_by_col(data_frame, experiment_design_file, control_label)
}
\arguments{
\item{data_frame}{a data frame as input.}

\item{experiment_design_file}{a data frame for design of experiment.}

\item{control_label}{a string for a control.}
}
\value{
A data frame after normalization.
}
\description{
For data without pairs but with control, normalize them to the control.
}
