% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_nopair_noctrl_by_colmed.R
\name{normalize_nopair_noctrl_by_colmed}
\alias{normalize_nopair_noctrl_by_colmed}
\title{For data without pairs and control, normalize them to the median.}
\usage{
normalize_nopair_noctrl_by_colmed(data_frame)
}
\arguments{
\item{data_frame}{a data frame as input.}
}
\value{
A data frame after normalization.
}
\description{
For data without pairs and control, normalize them to the median.
}
