% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_to_column_control_noPair_Control.R
\name{normalize_to_column_control_noPair_Control}
\alias{normalize_to_column_control_noPair_Control}
\title{For data without pairs but with control, normalize them to the control.}
\usage{
normalize_to_column_control_noPair_Control(data_frame, experiment_design_file,
  control_label)
}
\arguments{
\item{data_frame}{a data frame as input.}

\item{experiment_design_file}{a data frame for design of experiment.}

\item{control_label}{a string for a control.}
}
\value{
A data frame after normalization.
}
\description{
For data without pairs but with control, normalize them to the control.
}
