% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_to_column_median_noPair_noControl.R
\name{normalize_to_column_median_noPair_noControl}
\alias{normalize_to_column_median_noPair_noControl}
\title{For data without pairs and control, normalize them to the median.}
\usage{
normalize_to_column_median_noPair_noControl(data_frame)
}
\arguments{
\item{data_frame}{a data frame as input.}
}
\value{
A data frame after normalization.
}
\description{
For data without pairs and control, normalize them to the median.
}
