% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seqlogo.R
\name{plot_seqlogo}
\alias{plot_seqlogo}
\title{Plot sequence logo based on list that consist of motifs and sequences.}
\usage{
plot_seqlogo(base_dir, foreground_sequences_mapped_to_motifs,
  plot_min_seqs = 5)
}
\arguments{
\item{base_dir}{A path used for saving plots.}

\item{foreground_sequences_mapped_to_motifs}{A list that consist of motifs and sequences.}

\item{plot_min_seqs}{A numeric value for cutoff, sequences of motifs greater than the cutoff are plotted, the default is 5.}
}
\value{
Plot sequence logo based on list that consist of motifs and sequences.
The results will be saved in a folder named PhosMap_ggseqlogo in the BASE_DIR
parameter specified directory.
}
\description{
Plot sequence logo based on list that consist of motifs and sequences.
}
\examples{
ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/foreground_sequences_mapped_to_motifs.RData"
load_data <- load_data_with_ftp(ftp_url, 'RData')
writeBin(load_data, "foreground_sequences_mapped_to_motifs.RData")
load("foreground_sequences_mapped_to_motifs.RData")

BASE_DIR = getwd() # current working directory
BASE_DIR = normalizePath(BASE_DIR)
plot_seqlogo(BASE_DIR, foreground_sequences_mapped_to_motifs, plot_min_seqs = 25)

}
\references{
(1) Omar Wagih (2017). ggseqlogo: A 'ggplot2' Extension for Drawing Publication-Ready Sequence Logos. R package version 0.1.\
https://github.com/omarwagih/ggseqlogo; (2) Hadley Wickham (2018). stringr: Simple, Consistent Wrappers for Common String Operations. \
R package version 1.3.0. https://CRAN.R-project.org/package=stringr
}
\author{
Dongdong Zhan and Mengsha Tong
}
