% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_filter_psites.R
\name{pre_process_filter_psites}
\alias{pre_process_filter_psites}
\title{Get peptides data frame passed phosphorylation sites quality control.}
\usage{
pre_process_filter_psites(firmiana_peptide_dir, psites_score_dir,
  phospho_experiment_design_file_path, qc, min_score = 20, min_FDR = 0.01)
}
\arguments{
\item{firmiana_peptide_dir}{A folder containing peptide identification files from Firmiana as input.}

\item{psites_score_dir}{A folder containing psites scores files extracted from mascot xml as input.}

\item{phospho_experiment_design_file_path}{A string representing the path of phospho-experiment design file as input.}

\item{qc}{A boolean value representing whether it has QC files. The default is True.}

\item{min_score}{A numeric for the minimum score of credible peptides, the default is 20 for Mascot ion score.}

\item{min_FDR}{A numeric for the minimum FDR of credible peptides, the default is 0.01.}
}
\value{
A merged data frame containing sequence, gi, psite, area and psm.
}
\description{
Filter phosphorylation sites by extracting all peptides with ion score>=20 and FDR<0.01 from Firmiana and having psites scores.
Generate new IDs consisting of sequence, gi, psite.
Quantification values containing area and psm.
}
\examples{
\dontrun{
merge_df_with_phospho_peptides = pre_process_filter_psites(
  firmiana_peptide_dir,
  psites_score_dir
)
}
}
