% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_maxquant.R
\name{qc_maxquant}
\alias{qc_maxquant}
\title{Quality control for phosphoproteomics data from MaxQuant.}
\usage{
qc_maxquant(
  data_frame,
  experiment_code_file_path,
  min_score = 40,
  min_loc_prob = 0.75,
  max_na_num = 2
)
}
\arguments{
\item{data_frame}{A data frame of Phospho (STY)Sites.txt.}

\item{experiment_code_file_path}{Experiment code file path.}

\item{min_score}{Threshold for MaxQuant score. The default is 40.}

\item{min_loc_prob}{Threshold for MaxQuant Localization.prob. The default is 0.75.}

\item{max_na_num}{Threshold for the number of missing values. The default is 2.}
}
\value{
A result list. Elements are a data frame containing quality control result and a data frame containing information required for motif analysis.
}
\description{
Quality control for phosphoproteomics data from MaxQuant.
}
\examples{
\dontrun{
rawdata <- read.csv("Phospho (STY)Sites.txt",header=T,sep='\t')
qc_results <- qc_maxquant(rawdata, "./experiment_code_file.txt", min_score = 40, min_loc_prob = 0.75, max_na_num = 2)
qc_result <- qc_results[[1]]
qc_result_for_motifanalysis <- qc_results[[2]]
}
}
