% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seach_motif_pattern.R
\name{seach_motif_pattern}
\alias{seach_motif_pattern}
\title{Convert data frame of motif to the sequence pattern}
\usage{
seach_motif_pattern(
  foreground_sequence,
  background_sequence,
  min_sequence_count = 1,
  min_pvalue = 0.01,
  center = "S",
  width
)
}
\arguments{
\item{foreground_sequence}{A vector for AA sequences with fixed length as foreground input.}

\item{background_sequence}{A vector for AA sequences with fixed length as background input.}

\item{min_sequence_count}{A numeric for the minimum sequence number assigned to a motif.}

\item{min_pvalue}{A numeric for the minimum pvalue for found motif.}

\item{center}{A character for center of k-mer.}

\item{width}{A numeric for specific k-mer.}
}
\value{
A list for information summary of searching mortif
}
\description{
Convert data frame of motif to the sequence pattern
}
\examples{
## The process needs to load data from PhosMap datasets stored into FTP server and perform large computation.
## It may take a few minutes.
if(FALSE){
ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/seach_motif_pattern.RData"
load_data <- load_data_with_ftp(ftp_url, 'RData')
writeBin(load_data, "seach_motif_pattern.RData")
load("seach_motif_pattern.RData")

motif_result_loop_i <- seach_motif_pattern(
  loop_foreground[1:100],
  loop_background[1:1000],
  min_sequence_count = min_sequence_count,
  min_pvalue = min_pvalue,
  center = center,
  width = check_result_list$width
)
head(motif_result_loop_i)

}
}
\references{
Omar Wagih (2014). rmotifx: An iterative statistical approach to the discovery of biological sequence motifs. R package version 1.0.
}
\author{
Dongdong Zhan and Mengsha Tong
}
