% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_with_umap.R
\name{visualization_with_umap}
\alias{visualization_with_umap}
\title{A umap plot.}
\usage{
visualization_with_umap(
  expr_data_frame,
  group,
  main = "UMAP",
  n_neighbors = 10
)
}
\arguments{
\item{expr_data_frame}{A data frame containing ID and quantification value.}

\item{group}{A factor for group information.}

\item{main}{The main title of plot.}

\item{n_neighbors}{A numerical value for the size of local neighborhood, the default is 10.}
}
\value{
A umap plot.
}
\description{
A umap plot.
}
