% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csv_pep_seq_conf.R
\name{write_csv_pep_seq_conf}
\alias{write_csv_pep_seq_conf}
\title{Write data to specific direction with CSV format.}
\usage{
write_csv_pep_seq_conf(expName, outputName, mascotfileNames, refFileName)
}
\arguments{
\item{expName}{a string for experiment name as input.}

\item{outputName}{a string for experiment name as output.}

\item{mascotfileNames}{a vector for storing mascot file names.}

\item{refFileName}{a string for reference file name.}
}
\description{
Write data to specific direction with CSV format.
}
\author{
Dongdong Zhan and Mengsha Tong
}
