% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_deps_sam.R
\name{analysis_deps_sam}
\alias{analysis_deps_sam}
\title{Differential expression analysis using SAM}
\usage{
analysis_deps_sam(expr_data_frame, group, log2_label = FALSE, nperms = 100,
  rand = NULL, minFDR = 0.05, samr_plot = TRUE)
}
\arguments{
\item{expr_data_frame}{A data frame containing ID and quantification value.}

\item{group}{A factor representing groups.}

\item{log2_label}{A boolean value for representing whether or not the value is logarithmic, the default is FALSE.}

\item{nperms}{Number of permutations used to estimate false discovery rates.}

\item{rand}{if specified, the random number generator will be put in a reproducible state.}

\item{minFDR}{A numeric value for filtering significant genes, the default is 0.05.}

\item{samr_plot}{A boolean value for representing whether or not samr graph is plotted.}
}
\value{
A list containing results from sam analysis.
}
\description{
Differential expression analysis using SAM
}
\examples{
\dontrun{
sam_results_list = analysis_deps_sam(
  expr_data_frame, group, log2_label = FALSE,
  nperms = 100, rand = NULL, minFDR = 0.05,samr_plot = T
)
}
}
\references{
R. Tibshirani, G. Chu, T. Hastie and Balasubramanian Narasimhan (2010). samr: SAM: Significance Analysis of Microarrays.\
Rpackage version 1.28. https://CRAN.R-project.org/package=samr
}
\author{
Dongdong Zhan and Mengsha Tong
}
