% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mea_input.R
\name{check_mea_input}
\alias{check_mea_input}
\title{Check input for motif enrichment analysis (mea)}
\usage{
check_mea_input(foreground, background, center)
}
\arguments{
\item{foreground}{A vector for AA sequences with fixed length as foreground input.}

\item{background}{A vector for AA sequences with fixed length as background input.}

\item{center}{A character for center of k-mer.}
}
\value{
A list passing check steps
}
\description{
Check input for motif enrichment analysis (mea)
}
\examples{
\dontrun{
check_result_list <- check_mea_input(
  foreground,
  background,
  center
)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
