% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_kses.R
\name{compute_kses}
\alias{compute_kses}
\title{computing kinase-substrate enrichment score}
\usage{
compute_kses(substate_vector, regulons_of_kinase,
  substrates_of_kinase_in_exp_count)
}
\arguments{
\item{substate_vector}{a vector for substrates with values indentified in current experiments.}

\item{regulons_of_kinase}{a vector for substrates of a specific kinase, which with substrates identified in current experiments.}

\item{substrates_of_kinase_in_exp_count}{a numeric for numbers in regulons_of_kinase vector.}
}
\value{
A numeric or NA for enrichment_score.
}
\description{
computing kinase-substrate enrichment score
}
\examples{
\dontrun{
enrichment_score <- compute_kses(
  substate_vector,
  regulons_of_kinase,
  substrates_of_kinase_in_exp_count
)
}

}
\references{
Hernandez-Armenta C et al. Benchmarking substrate-based kinase activity inference using phosphoproteomic data[J]. Bioinformatics, 2017, 33(12): 1845-1851.
}
\author{
Dongdong Zhan and Mengsha Tong
}
