% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_pwm.R
\name{construct_pwm}
\alias{construct_pwm}
\title{Construct position weight matrix}
\usage{
construct_pwm(sequences, width, frequency_flag = TRUE)
}
\arguments{
\item{sequences}{A vector for aligned sequences with fixed length.}

\item{width}{A numeric for specific k-mer.}

\item{frequency_flag}{A boolean for showing real frequency or frequency probability, the default is TRUE for showing real frequency.}
}
\value{
A position weight matrix.
}
\description{
Construct position weight matrix
}
\examples{
\dontrun{
pwm <- construct_pwm(
  sequences,
  width,
  frequency_flag = TRUE
)
}

}
\references{
Wagih O, Sugiyama N, Ishihama Y, et al. Uncovering phosphorylation-based specificities through functional interaction networks[J]. Molecular & Cellular Proteomics, 2016, 15(1): 236-245.
}
\author{
Dongdong Zhan and Mengsha Tong
}
