% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fore_seq_to_motif.R
\name{fore_seq_to_motif}
\alias{fore_seq_to_motif}
\title{Convert the list that consists of motifs and the corresponding sequences to data frame.}
\usage{
fore_seq_to_motif(foreground_sequences_mapped_to_motifs)
}
\arguments{
\item{foreground_sequences_mapped_to_motifs}{A list that consists of motifs and the corresponding sequences.}
}
\value{
A data frame that consist of aligned sequences and the corresponding motifs.
}
\description{
Convert the list that consists of motifs and the corresponding sequences to data frame.
}
\examples{
\dontrun{
df <- fore_seq_to_motif(
  foreground_sequences_mapped_to_motifs
)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
