% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatted_output_mef_results.R
\name{formatted_output_mef_results}
\alias{formatted_output_mef_results}
\title{Output formatted sequences in foreground that are mapped to specific motif.}
\usage{
formatted_output_mef_results(foreground_sequences_mapped_to_motifs)
}
\arguments{
\item{foreground_sequences_mapped_to_motifs}{A list that consists of motifs and their corresponding aligned sequences from foreground.}
}
\value{
A data frame that motifs and their corresponding aligned sequences from foreground.
}
\description{
Output formatted sequences in foreground that are mapped to specific motif.
}
\examples{
\dontrun{
formatted_output_df <- formatted_output_mef_results(
  foreground_sequences_mapped_to_motifs
)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
