% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_combined_data_frame.R
\name{get_combined_data_frame}
\alias{get_combined_data_frame}
\title{Get a data frame mapped ID to Gene Symbol.}
\usage{
get_combined_data_frame(merge_df_with_phospho_peptides, species = "human",
  id_type = "RefSeq_Protein_GI")
}
\arguments{
\item{merge_df_with_phospho_peptides}{A dataframe consisting of IDs (Sequence_GI_Psite) and Area values.}

\item{species}{A string, the options are human, mouse and rattus, the default is human.}

\item{id_type}{A string, the options are 'GeneID', 'RefSeq_Protein_GI', 'RefSeq_Protein_Accession' and 'Uniprot_Protein_Accession', the default is RefSeq_Protein_GI.}
}
\value{
A dataframe with Sequence, GI, Modification, Gene Symbol, Area values and PSMs
}
\description{
This is an intermediate file and a dataframe with Gene Symbol is exported.
Based on a library file consisting of mapping relationships about Gene Symbol, GeneID, RefSeq_Protein_GI, RefSeq_Protein_Accession and Uniprot_Protein_Accession,
a new dataframe with Sequence, GI, Modification, Gene Symbol, Area and PSMs,is contructed.
}
\examples{
\dontrun{
combinated_df_with_mapped_gene_symbol <- get_combined_data_frame(
  merge_df_with_phospho_peptides
)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
