% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_with_AAs_i.R
\name{get_df_with_AAs_i}
\alias{get_df_with_AAs_i}
\title{Get a data frame of amino acid sequences for proteins.}
\usage{
get_df_with_AAs_i(unique_proteins, i, id_data_only_peptide2gi,
  proteins_in_id_data_only_peptide2gi, sequences_in_id_data_only_peptide2gi,
  modification_index_in_protein_seq_list)
}
\arguments{
\item{unique_proteins}{a vector for unique proteins.}

\item{i}{the ith unique proteins.}

\item{id_data_only_peptide2gi}{a data frame for peptides with protein gi.}

\item{proteins_in_id_data_only_peptide2gi}{a vector for proteins with only protein gi.}

\item{sequences_in_id_data_only_peptide2gi}{a vector for peptides with only protein gi.}

\item{modification_index_in_protein_seq_list}{a list for the index of modifications in protein sequence.}
}
\value{
A data frame with sequences for proteins.
}
\description{
Get a data frame of amino acid sequences for proteins.
}
\examples{
## The process needs to load data from PhosMap datasets stored into FTP server and perform large computation.
## It may take a few minutes.
if(FALSE){
ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/get_df_with_AAs_i.RData"
load_data <- load_data_with_ftp(ftp_url, 'RData')
writeBin(load_data, "get_df_with_AAs_i.RData")
load("get_df_with_AAs_i.RData")

df_with_AAs_i <- get_df_with_AAs_i(unique_proteins, i,
  id_data_only_peptide2gi,
  proteins_in_id_data_only_peptide2gi,
  sequences_in_id_data_only_peptide2gi,
  modification_index_in_protein_seq_list
)
head(df_with_AAs_i)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
