% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_with_AAs_i.R
\name{get_df_with_AAs_i}
\alias{get_df_with_AAs_i}
\title{Get data frame of amino acide sequences for a protein.}
\usage{
get_df_with_AAs_i(unique_proteins, i, id_data_only_peptide2gi,
  proteins_in_id_data_only_peptide2gi, sequences_in_id_data_only_peptide2gi,
  modification_index_in_protein_seq_list)
}
\arguments{
\item{unique_proteins}{a vector for unique proteins.}

\item{i}{the ith unique proteins.}

\item{id_data_only_peptide2gi}{a data frame for peptides with protein gi.}

\item{proteins_in_id_data_only_peptide2gi}{a vector for proteins with only protein gi.}

\item{sequences_in_id_data_only_peptide2gi}{a vector for peptides with only protein gi.}

\item{modification_index_in_protein_seq_list}{a list for the index of modifications in protein sequence.}
}
\value{
A data frame with sequences for a protein.
}
\description{
Get data frame of amino acide sequences for a protein.
}
\author{
Dongdong Zhan and Mengsha Tong
}
