% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_suffix.R
\name{get_file_suffix}
\alias{get_file_suffix}
\title{Get sufffix of input file.}
\usage{
get_file_suffix(file_name)
}
\arguments{
\item{file_name}{A string for file name.}
}
\value{
A string for file format.
}
\description{
Get sufffix of input file.
}
