% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filtered_df.R
\name{get_filtered_df}
\alias{get_filtered_df}
\title{Get filtered data frame based on the Mascot and reference file.}
\usage{
get_filtered_df(mascotfileName, refFileName)
}
\arguments{
\item{mascotfileName}{a string for mascot name as input.}

\item{refFileName}{a string for reference file name.}
}
\value{
A filtered data frame
}
\description{
Get filtered data frame based on the Mascot and reference file.
}
\author{
Dongdong Zhan and Mengsha Tong
}
