% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_foreground_seq_to_motifs.R
\name{get_foreground_seq_to_motifs}
\alias{get_foreground_seq_to_motifs}
\title{Get motifs and their corresponding aligned sequences form from foreground.}
\usage{
get_foreground_seq_to_motifs(motifs_list, foreground)
}
\arguments{
\item{motifs_list}{A list from motif enrichment analysis.}

\item{foreground}{A vector for aligned sequences.}
}
\value{
A list containing motifs and the corresponding sequences from foreground.
}
\description{
Get motifs and their corresponding aligned sequences form from foreground.
}
\examples{
\dontrun{
mots_match_list = get_foreground_seq_to_motifs(
  motifs_list,
  foreground
)
}
}
\references{
Hadley Wickham (2018). stringr: Simple, Consistent Wrappers for Common String Operations. R package version 1.3.0.\
https://CRAN.R-project.org/package=stringr.
}
\author{
Dongdong Zhan and Mengsha Tong
}
