% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ka_by_mean_or_mlr.R
\name{get_ka_by_mean_or_mlr}
\alias{get_ka_by_mean_or_mlr}
\title{Computing kinase activity using mean value and multiple linear regression (ridge regression) except KSEA}
\usage{
get_ka_by_mean_or_mlr(ptypes_data, species = "human", log2_label = FALSE,
  method = "mean")
}
\arguments{
\item{ptypes_data}{A data frame of phosphorylation data after normalization.}

\item{species}{A string representing the species of imported data, the options are human, mouse and rat.}

\item{log2_label}{A boolean value representing whether data is logarithmetics, the default is FALSE.}

\item{method}{A string for the method to compute kinase activity, the options are 'mean' and 'mlr' (multiple linear regression),
the default is mean.}
}
\value{
A data frame that consists of kinase, psite, substrate, counting byond ratio_cutoff and corresponding original value.
}
\description{
Computing kinase activity using mean value and multiple linear regression (ridge regression) except KSEA
}
\examples{
\dontrun{
kinase_activity_df <- get_ka_by_mean_or_mlr(
  ptypes_data,
  species = 'human',
  log2_label = TRUE,
  method = 'mean'
)
}



}
\author{
Dongdong Zhan and Mengsha Tong
}
