% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ksea_result_list.R
\name{get_ksea_result_list}
\alias{get_ksea_result_list}
\title{Kinase activity analysis based on known and predicted relationship of kinase-substrate}
\usage{
get_ksea_result_list(ptypes_data_ratio_in_sigle_exp, ID,
  kinase_substrate_regulation_relationship, ksea_activity_i_pvalue = 0.05)
}
\arguments{
\item{ptypes_data_ratio_in_sigle_exp}{A quantifcation vector from single experiment.}

\item{ID}{A phosporylation ID vector like VIM_S56 (GeneSymbol_psite).}

\item{kinase_substrate_regulation_relationship}{A data frame contanning relationship of kinase-substrate that consists of "kinase", "substrate", "site", "sequence" and "predicted" columns.}

\item{ksea_activity_i_pvalue}{A cutoff used for filtering significant activities computed from KSEA.}
}
\value{
A list containing results from ksea.
}
\description{
Kinase activity analysis based on known and predicted relationship of kinase-substrate
}
\examples{
\dontrun{
ksea_result_list = get_ksea_result_list(
  ptypes_data_ratio_in_sigle_exp,
  ID,
  kinase_substrate,
  0.05
)
}


}
\author{
Dongdong Zhan and Mengsha Tong
}
