% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modification_index.R
\name{get_modification_index}
\alias{get_modification_index}
\title{Get indexes of modifications in protein sequences.}
\usage{
get_modification_index(id_data_only_peptide2gi, fasta_data)
}
\arguments{
\item{id_data_only_peptide2gi}{a data frame for peptides with protein gi.}

\item{fasta_data}{a fasta data for a specific species.}
}
\value{
A vector for indexes of modifications in protein sequences.
}
\description{
Get indexes of modifications in protein sequences.
}
\author{
Dongdong Zhan and Mengsha Tong
}
