% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modification_index.R
\name{get_modification_index}
\alias{get_modification_index}
\title{Get indexes of modifications in protein sequences.}
\usage{
get_modification_index(id_data_only_peptide2gi, fasta_data)
}
\arguments{
\item{id_data_only_peptide2gi}{a data frame for peptides with protein gi.}

\item{fasta_data}{a fasta data for a specific species.}
}
\value{
A vector for indexes of modifications in protein sequences.
}
\description{
Get indexes of modifications in protein sequences.
}
\examples{
## The process needs to load data from PhosMap datasets stored into FTP server and perform large computation.
## It may take a few minutes.
if(FALSE){
ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/get_modification_index.RData"
load_data <- load_data_with_ftp(ftp_url, 'RData')
writeBin(load_data, "get_modification_index.RData")
load("get_modification_index.RData")

modification_index_in_protein_seq_list <- get_modification_index(
  id_data_only_peptide2gi[1:100, ],
  fasta_data
)
head(modification_index_in_protein_seq_list)
}

}
\author{
Dongdong Zhan and Mengsha Tong
}
