% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_normalized_data_FOT5.R
\name{get_normalized_data_FOT5}
\alias{get_normalized_data_FOT5}
\title{Normailization on basis of sum}
\usage{
get_normalized_data_FOT5(data_frame, experiment_code_file_path)
}
\arguments{
\item{data_frame}{A data frame containing IDs and values merged from multi-experiments as input.}

\item{experiment_code_file_path}{A file path of storing experiment codes as input. The experiment codes are required to keep pace with column names of Values.}
}
\value{
A data frame after normalization
}
\description{
Normailization on basis of sum
}
\examples{
\dontrun{
data_frame_normalization = get_normalized_data_FOT5(
  data_frame,
  experiment_code_file_path
)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
