% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_substrate_expr_df.R
\name{get_substrate_expr_df}
\alias{get_substrate_expr_df}
\title{Get a data frame only containing kinase inferred by KSEA}
\usage{
get_substrate_expr_df(ID, kinase_substrate_regulation_relationship,
  ksea_regulons, ptypes_data_ratio, ratio_cutoff = 3)
}
\arguments{
\item{ID}{A phosporylation ID vector like VIM_S56 (GeneSymbol_psite).}

\item{kinase_substrate_regulation_relationship}{A data frame contanning relationship of
kinase-substrate that consists of "kinase", "substrate", "site", "sequence" and "predicted" columns.}

\item{ksea_regulons}{A kinase vector from ksea}

\item{ptypes_data_ratio}{A data frame that the ratio of phosphorylation and profiling data}

\item{ratio_cutoff}{A cutoff that depicts quantification changes
at phosphorylation level relative to profiling level, the default is 3.}
}
\value{
A data frame that consists of kinase, psite, substrate,
counting byond ratio_cutoff and corresponding original value.
}
\description{
Get a data frame only containing kinase inferred by KSEA
}
\examples{
\dontrun{
kinase_site_substrate_original_ratio_df <- get_substrate_expr_df(
  ID,
  kinase_substrate_regulation_relationship,
  ksea_regulons,
  ptypes_data_ratio,
  ratio_cutoff = 3
)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
