% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_with_unique_sites.R
\name{get_summary_with_unique_sites}
\alias{get_summary_with_unique_sites}
\title{Assign psites to protein sequence.}
\usage{
get_summary_with_unique_sites(combined_df_with_mapped_gene_symbol,
  species = "human", fasta_type = "refseq")
}
\arguments{
\item{combined_df_with_mapped_gene_symbol}{A dataframe with Sequence, ID, Modification, Gene Symbol, Area and PSMs as input.}

\item{species}{A string, the options are human, mouse and rattus, the default is human.}

\item{fasta_type, }{A string for fasta source, the options are refseq and uniprot, the default is refseq}
}
\value{
A dataframe that all redundant psites are assigned to protein sequence.
}
\description{
Construct the data frame with unique phosphorylation site for each protein sequence and eliminate redundancy.
}
\examples{
## The process needs to load data from PhosMap datasets stored into FTP server and perform large computation.
## It may take a few minutes.
if(FALSE){
ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/get_summary_with_unique_sites.RData"
load_data <- load_data_with_ftp(ftp_url, 'RData')
writeBin(load_data, "get_summary_with_unique_sites.RData")
load("get_summary_with_unique_sites.RData")

summary_df_of_unique_proteins_with_sites <- get_summary_with_unique_sites(
  combined_df_with_mapped_gene_symbol[1:100, ],
  species = 'human',
  fasta_type = 'refseq'
)
head(summary_df_of_unique_proteins_with_sites)

}
}
\author{
Dongdong Zhan and Mengsha Tong
}
