% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_AAs_i_df.R
\name{get_unique_AAs_i_df}
\alias{get_unique_AAs_i_df}
\title{Get data frame without redundancy.}
\usage{
get_unique_AAs_i_df(df_with_AAs_i)
}
\arguments{
\item{df_with_AAs_i}{a data frame for peptides of the ith protein.}
}
\value{
A data frame with sites in unique protein.
}
\description{
Get data frame without redundancy.
}
\author{
Dongdong Zhan and Mengsha Tong
}
