% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_psites_with_max_in_topX.R
\name{keep_psites_with_max_in_topX}
\alias{keep_psites_with_max_in_topX}
\title{Keep psites whose row maximum is top N (percentage).}
\usage{
keep_psites_with_max_in_topX(phospho_data, percent_of_kept_sites = 3/4)
}
\arguments{
\item{phospho_data}{A data frame of phospho-data.}

\item{percent_of_kept_sites}{A numeric value representing a cutoff used for filter psites. The default is 3/4.}
}
\value{
A data frame meeting specific cutoff.
}
\description{
Compute row maximum each psites, sort row maximum in decreasing order and keep top N (percentage).
}
\examples{
ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/phospho_data.RData"
load_data <- load_data_with_ftp(ftp_url, 'RData')
writeBin(load_data, "phospho_data.RData")
load("phospho_data.RData")

phospho_data_topX <- keep_psites_with_max_in_topX(phospho_data,
  percent_of_kept_sites = 0.9)

head(phospho_data_topX)

}
\author{
Dongdong Zhan and Mengsha Tong
}
