% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_with_http.R
\name{load_data_with_http}
\alias{load_data_with_http}
\title{Load datasets from URL (https://github.com/ecnuzdd/PhosMap_datasets)}
\usage{
load_data_with_http(http_link, data_type)
}
\arguments{
\item{http_link}{A string for URL of datasets.}

\item{data_type}{A string for type of datasets (txt of csv).}
}
\value{
A dataframe
}
\description{
Some datasets with larger size need to be loaded for mapping ID or protein sequence when using PhosMap.
These datasets could be ragarded as library and uploaded to https://github.com/ecnuzdd/PhosMap_datasets in advance.
When first perfoming functions depending on these datasets, PhosMap will get them from specific URL and save them into local disk.
}
\examples{
\dontrun{
http_link <- url('https://raw.githubusercontent.com/ecnuzdd/PhosMap_datasets/master/function_demo_data/profiling_exp_design_info.txt')
data_type <- 'txt'
load_data <- load_data_with_http(
  http_link, data_type
)
head(load_data)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
