% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mea_based_on_background.R
\name{mea_based_on_background}
\alias{mea_based_on_background}
\title{Motif enrichment based on global background (fasta library from Refseq).}
\usage{
mea_based_on_background(foreground, AA_in_protein, background, motifx_pvalue)
}
\arguments{
\item{foreground}{A vector for aligned sequence of foreground.}

\item{AA_in_protein}{A vector for the location of S/T/Y in sequence of protein.}

\item{background}{A vector for aligned sequence of background.}

\item{motifx_pvalue}{A numeric value for selecting motifs that meets the minimum cutoff.}
}
\value{
A list containing motifs and the corresponding sequences
}
\description{
Motif enrichment based on global background (fasta library from Refseq).
}
\examples{
demo_data_url <- url('https://raw.githubusercontent.com/ecnuzdd/PhosMap_datasets/master/function_demo_data/foreground.RData')
load(demo_data_url)

motifs_list <- mea_based_on_background(
  foreground,
  AA_in_protein,
  background,
  motifx_pvalue
)
head(motifs_list)

}
\author{
Dongdong Zhan and Mengsha Tong
}
