% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_data_frame_to_sequence.R
\name{motif_data_frame_to_sequence}
\alias{motif_data_frame_to_sequence}
\title{Convert data frame of motif to the sequence pattern}
\usage{
motif_data_frame_to_sequence(motif_data_frame, center, width)
}
\arguments{
\item{motif_data_frame}{A data frame with two columns including amino acid and index on sequence with fixed length.}

\item{center}{A character for center of k-mer.}

\item{width}{A numeric for specific k-mer.}
}
\value{
A string for motif pattern
}
\description{
Convert data frame of motif to the sequence pattern
}
\examples{
\dontrun{
motif_pattern <- motif_data_frame_to_sequence(
  motif_data_frame,
  center,
  width
)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
