% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_nopair_noctrl_by_colmed.R
\name{normalize_nopair_noctrl_by_colmed}
\alias{normalize_nopair_noctrl_by_colmed}
\title{For data without pairs and control, normalize them to the median.}
\usage{
normalize_nopair_noctrl_by_colmed(data_frame)
}
\arguments{
\item{data_frame}{a data frame as input.}
}
\value{
A data frame after normalization.
}
\description{
For data without pairs and control, normalize them to the median.
}
\examples{
## The process needs to load data from PhosMap datasets stored into FTP server and perform large computation.
## It may take a few minutes.
if(FALSE){
ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/normalize_nopair_noctrl_by_colmed.RData"
load_data <- load_data_with_ftp(ftp_url, 'RData')
writeBin(load_data, "normalize_nopair_noctrl_by_colmed.RData")
load("normalize_nopair_noctrl_by_colmed.RData")

phospho_data_normalize_by_column <- normalize_nopair_noctrl_by_colmed(
  phospho_data_normalized
)
head(phospho_data_normalize_by_column)
}
}
