% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_phos_data_to_profiling.R
\name{normalize_phos_data_to_profiling}
\alias{normalize_phos_data_to_profiling}
\title{Normalize phospho-data to profiling}
\usage{
normalize_phos_data_to_profiling(phospho_data_normalized,
  profiling_data_normalized, phosphorylation_exp_design_info_file_path,
  profiling_exp_design_info_file_path, control_label = NA,
  pair_flag = FALSE)
}
\arguments{
\item{phospho_data_normalized}{A data frame of phospho-data after normalization}

\item{profiling_data_normalized}{A data frame of profiling after normalization}

\item{phosphorylation_exp_design_info_file_path}{A file path about phosphorylation experiment design, it has 2 kinds of file configuration as follows:
1. experiment_design_noPair.txt must contain columns of Experiment_Code, Group.
2. experiment_design_Pair.txt must contain columns of Experiment_Code, Group, and Pair. (Pair: 1 -> case, -1 -> control)}

\item{profiling_exp_design_info_file_path}{A file path about profiling experiment design, it has 2 kinds of file configuration as same as phosphorylation_exp_design_info_file_path.}

\item{control_label}{A string represents label of control group. The default is NA which shows no control group.}

\item{pair_flag}{A boolean value represents whether experiments have pairs. The default is FALSE which shows no pairs.}
}
\value{
A data frame which comes from results that phospho-data is normalizated base on the abundance of proteins in the profiling experiments.
}
\description{
Normalize phospho-data to profiling
}
\examples{
ftp_url1 <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/phospho_data_topX.RData"
ftp_url2 <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/profiling_data_normalized.RData"
ftp_url3 <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/phosphorylation_exp_design_info.txt"
ftp_url4 <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/profiling_exp_design_info.txt"

load_data1 <- load_data_with_ftp(ftp_url1, 'Rdata')
writeBin(load_data1, "phospho_data_topX.RData")
load("phospho_data_topX.RData")

load_data2 <- load_data_with_ftp(ftp_url2, 'Rdata')
writeBin(load_data2, "profiling_data_normalized.RData")
load("profiling_data_normalized.RData")

load_data3 <- load_data_with_ftp(ftp_url3, 'downloadtxt')
writeBin(load_data3, "phosphorylation_exp_design_info.txt")
phosphorylation_exp_design_info_file_path <- "./phosphorylation_exp_design_info.txt"

load_data4 <- load_data_with_ftp(ftp_url4, 'downloadtxt')
writeBin(load_data4, "profiling_exp_design_info.txt")
profiling_exp_design_info_file_path <- "./profiling_exp_design_info.txt"

data_frame_normalization_with_control_no_pair <- normalize_phos_data_to_profiling(
  phospho_data_topX, profiling_data_normalized,
  phosphorylation_exp_design_info_file_path,
  profiling_exp_design_info_file_path,
  control_label = '0',
  pair_flag = FALSE
)
head(data_frame_normalization_with_control_no_pair)

}
