% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_to_Pair.R
\name{normalize_to_Pair}
\alias{normalize_to_Pair}
\title{For data with pairs, normalize them to the sample with flag eaqul to -1.}
\usage{
normalize_to_Pair(data_frame, experiment_design_file)
}
\arguments{
\item{data_frame}{a data frame as input.}

\item{experiment_design_file}{a data frame for design of experiment.}
}
\value{
A data frame after normalization.
}
\description{
For data with pairs, normalize them to the sample with flag eaqul to -1.
}
\examples{
## The process needs to load data from PhosMap datasets stored into FTP server and perform large computation.
## It may take a few minutes.
if(FALSE){
ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/normalize_to_Pair.RData"
load_data <- load_data_with_ftp(ftp_url, 'Rdata')
writeBin(load_data, "normalize_to_Pair.RData")
load("normalize_to_Pair.RData")

phospho_data_normalize_by_column <- normalize_to_Pair(
  phospho_data_normalized,
  pairing_phosphorylation_experiment_design_file
)
head(phospho_data_normalize_by_column)

}
}
