% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_to_Pair.R
\name{normalize_to_Pair}
\alias{normalize_to_Pair}
\title{For data with pairs, normalize them to the sample with flag eaqul to -1.}
\usage{
normalize_to_Pair(data_frame, experiment_design_file)
}
\arguments{
\item{data_frame}{a data frame as input.}

\item{experiment_design_file}{a data frame for design of experiment.}
}
\value{
A data frame after normalization.
}
\description{
For data with pairs, normalize them to the sample with flag eaqul to -1.
}
