% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seqlogo.R
\name{plot_seqlogo}
\alias{plot_seqlogo}
\title{Plot sequence logo based on list that consist of motifs and sequences.}
\usage{
plot_seqlogo(base_dir, foreground_sequences_mapped_to_motifs,
  plot_min_seqs = 5)
}
\arguments{
\item{base_dir}{A path used for saving plots.}

\item{foreground_sequences_mapped_to_motifs}{A list that consist of motifs and sequences.}

\item{plot_min_seqs}{A numeric value for cutoff, sequences of motifs greater than the cutoff are plotted, the default is 5.}
}
\description{
Plot sequence logo based on list that consist of motifs and sequences.
}
\examples{
\dontrun{
plot_seqlogo(base_dir, foreground_sequences_mapped_to_motifs, plot_min_seqs = 50)
}
}
\references{
(1) Omar Wagih (2017). ggseqlogo: A 'ggplot2' Extension for Drawing Publication-Ready Sequence Logos. R package version 0.1.\
https://github.com/omarwagih/ggseqlogo; (2) Hadley Wickham (2018). stringr: Simple, Consistent Wrappers for Common String Operations. \
R package version 1.3.0. https://CRAN.R-project.org/package=stringr
}
\author{
Dongdong Zhan and Mengsha Tong
}
