% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seach_motif_pattern.R
\name{seach_motif_pattern}
\alias{seach_motif_pattern}
\title{Convert data frame of motif to the sequence pattern}
\usage{
seach_motif_pattern(foreground_sequence, background_sequence,
  min_sequence_count = 1, min_pvalue = 0.01, center = "S", width)
}
\arguments{
\item{foreground_sequence}{A vector for AA sequences with fixed length as foreground input.}

\item{background_sequence}{A vector for AA sequences with fixed length as background input.}

\item{min_sequence_count}{A numeric for the minimum sequence number assigned to a motif.}

\item{min_pvalue}{A numeric for the minimum pvalue for found motif.}

\item{center}{A character for center of k-mer.}

\item{width}{A numeric for specific k-mer.}
}
\value{
A list for information summary of searching mortif
}
\description{
Convert data frame of motif to the sequence pattern
}
\examples{
\dontrun{
seach_motif_pattern(
    foreground_sequence,,
    background_sequence,
    min_sequence_count = 1,
    min_pvalue = 0.01,
    center = 'S',
    width = 15
)
}
}
\references{
Omar Wagih (2014). rmotifx: An iterative statistical approach to the discovery of biological sequence motifs. R package version 1.0.
}
\author{
Dongdong Zhan and Mengsha Tong
}
