% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_fuzzycluster.R
\name{visualization_fuzzycluster}
\alias{visualization_fuzzycluster}
\title{Visualize results from fuzzy clusters with line chart}
\usage{
visualization_fuzzycluster(input_data, group, group_levels, k_cluster,
  iteration = 100, mfrow = c(3, 3), min_mem = 0.1, plot = TRUE)
}
\arguments{
\item{input_data}{a data frame containing ID and expression profile.}

\item{group}{a factor for representing groups.}

\item{group_levels}{a factor levels for group.}

\item{k_cluster}{number of clusters fuzzy cluster.}

\item{iteration}{a numeric value for interation, the defualt is 100.}

\item{mfrow}{a vector containing 2 elements for controling the subplots in graphic window, the default is mfrow = c(3,3)}

\item{min_mem}{cutoff value for membership. Only results with greater than min_mem are showed.}

\item{plot}{a boolean value for deciding whether ploting, the default is TRUE.}
}
\value{
A lines chart with fuzzy degree.
}
\description{
Visualize results from fuzzy clusters with line chart
}
\examples{
ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/visualization_fuzzycluster.RData"
load_data <- load_data_with_ftp(ftp_url, 'RData')
writeBin(load_data, "visualization_fuzzycluster.RData")
load("visualization_fuzzycluster.RData")

fuzzy_clustObj <- visualization_fuzzycluster(
  fuzzy_input_df, group, group_levels,
  k_cluster=9, iteration = 100,
  mfrow = c(3,3), min_mem = 0.1
 )

}
\references{
(1) David Meyer, Evgenia Dimitriadou, Kurt Hornik, Andreas Weingessel and Friedrich Leisch (2017). e1071: Misc Functions of the \
Department of Statistics, Probability Theory Group (Formerly: E1071), TU Wien. R package version 1.6-8.https://CRAN.R-project.org/package=e1071 \
(2) Pengyi Yang (2018). ClueR: Cluster Evaluation. R package version 1.4. https://CRAN.R-project.org/package=ClueR
}
\author{
Dongdong Zhan and Mengsha Tong
}
