% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_with_simple_pca.R
\name{visualization_with_simple_pca}
\alias{visualization_with_simple_pca}
\title{A simple PCA plot.}
\usage{
visualization_with_simple_pca(expr_data_frame, main = "Simple PCA",
  point_cex = 2, point_col = "firebrick", point_type = 20, text_cex = 1)
}
\arguments{
\item{expr_data_frame}{A data frame containing ID and quantification value.}

\item{main}{The main title of plot.}

\item{point_cex}{a numerical value for point size.}

\item{point_col}{a color code or name for point color.}

\item{point_type}{point type, see points.}

\item{text_cex}{a numerical value for text size.}
}
\description{
A simple PCA plot.
}
\examples{
\dontrun{
visualization_with_simple_pca(expr_data_frame, main = 'Simple PCA',
point_cex = 2, point_col = 'firebrick', point_type = 20, text_cex = 1)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
