% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_with_simple_tsne.R
\name{visualization_with_simple_tsne}
\alias{visualization_with_simple_tsne}
\title{A simple t-SNE plot.}
\usage{
visualization_with_simple_tsne(expr_data_frame, group,
  main = "Simple t-SNE", perplexity = 10)
}
\arguments{
\item{expr_data_frame}{A data frame containing ID and quantification value.}

\item{group}{A factor for group information.}

\item{main}{The main title of plot.}

\item{perplexity}{A numerical value for perplexity, the default is 10.}
}
\value{
A simple t-SNE plot.
}
\description{
A simple t-SNE plot.
}
\examples{
ftp_url <- "ftp://111.198.139.72:4000/pub/PhosMap_datasets/function_demo_data/visualization_with_simple_tsne.RData"
load_data <- load_data_with_ftp(ftp_url, 'RData')
writeBin(load_data, "visualization_with_simple_tsne.RData")
load("visualization_with_simple_tsne.RData")

visualization_with_simple_tsne(
  expr_data_frame,
  group,
  main = 'Simple t-SNE',
  perplexity = 12
)

}
\author{
Dongdong Zhan and Mengsha Tong
}
