% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_with_simple_tsne.R
\name{visualization_with_simple_tsne}
\alias{visualization_with_simple_tsne}
\title{A simple t-SNE plot.}
\usage{
visualization_with_simple_tsne(expr_data_frame, group, main = "Simple t-SNE",
  perplexity = 10)
}
\arguments{
\item{expr_data_frame}{A data frame containing ID and quantification value.}

\item{group}{A factor for group information.}

\item{main}{The main title of plot.}

\item{perplexity}{A numerical value for perplexity, the default is 10.}
}
\description{
A simple t-SNE plot.
}
\examples{
\dontrun{
visualization_with_simple_tsne(
  expr_data_frame,
  group,
  main = 'Simple t-SNE',
  perplexity = 10
)
}
}
\author{
Dongdong Zhan and Mengsha Tong
}
